+++ CHANGELOG.md
@@ -547,7 +547,7 @@ setBetterPlayerGlobalKey.
 
 ## 0.0.23
 * General bug fixes.
-* Added player visibilityChangedBehavior in BetterPlayerConfiguration.
+* Added playerVisibilityChangedBehavior in BetterPlayerConfiguration.
 * Changed player behavior when player is not visible in viewport: if player was playing before leaving viewport it will be paused and if user views player again it will start playing automatically.
 * Added BetterPlayer.network and BetterPlayer.file methods.
 * Changed iOS & Android native classes name to prevent conflict issues with video_player.